requirements.txt
streamlit==1.42.1
python-dotenv==1.0.1
